requirements.txt
aniso8601==9.0.1
asgiref==3.4.1
awsebcli==3.20.3
blinker==1.4
boto3==1.24.0
botocore==1.27.96
brotlipy==0.7.0
cement==2.8.2
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.3
conda==4.10.3
coreapi==2.3.3
coreschema==0.0.4
cycler==0.11.0
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.6
Django==3.2.9
django-admin-list-filter-dropdown==1.0.3
django-allow-cidr==0.6.0
django-cors-headers==3.10.0
django-debug-toolbar==4.0.0
django-environ==0.9.0
django-extensions==3.1.5
django-filter==21.1
django-import-export==3.0.1
django-json-widget==1.1.1
django-rest-authtoken==2.1.4
django-safedelete==1.1.0
django-simple-history==3.0.0
django-soft-delete==0.9.21
django-storages==1.13.1
djangorestframework==3.12.4
djangorestframework-datatables==0.6.0
djangorestframework-simplejwt==5.0.0
drf-flex-fields==0.9.5
drf-yasg==1.20.0
et-xmlfile==1.1.0
filelock==3.8.0
Flask==2.1.2
flask-requests==0.0.14
Flask-RESTful==0.3.9
future==0.16.0
fuzzywuzzy==0.18.0
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.4
inflection==0.5.1
instagrapi==1.16.31
itsdangerous==2.1.2
itypes==1.2.0
Jinja2==3.1.2
jmespath==0.10.0
kiwisolver==1.3.2
Markdown==3.3.6
MarkupPy==1.14
MarkupSafe==2.1.1
matplotlib==3.4.3
menuinst==1.4.16
mysql-connector==2.2.9
numpy==1.21.3
odfpy==1.4.1
openai==0.11.5
openpyxl==3.0.9
packaging==21.2
pampy==0.3.0
pandas==1.3.4
pandas-stubs==1.2.0.40
pathspec==0.9.0
Pillow==8.4.0
platformdirs==2.5.3
psycopg2==2.9.2
pycosat==0.6.3
pycryptodomex==3.15.0
pydantic==1.10.2
PyJWT==2.3.0
pyodbc==4.0.32
pyparsing==2.4.7
pypiwin32==223
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2022.1
pywin32==228
PyYAML==5.4.1
requests==2.26.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
s3transfer==0.6.0
scipy==1.7.2
seaborn==0.11.2
semantic-version==2.8.5
six==1.14.0
sqlparse==0.4.2
tablib==3.2.1
termcolor==1.1.0
typing-extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.9
virtualenv==20.16.6
wcwidth==0.1.9
Werkzeug==2.1.2
wincertstore==0.2
xlrd==2.0.1
xlwt==1.3.0
zipp==3.8.0
